UnsupportedOperationException -> SparkUnsupportedOperationException
